Working 64bit

+++ src/io.c
@@ -282,7 +282,7 @@ static any rdList(void) {
 static any anonymous(any s) {
    int c, i;
    word w;
-   unsigned long n;
+   unsigned long long n;
    heap *h;
 
    if ((c = getByte1(&i, &w, &s)) != '$')
